// src/features/pages/resume/resume.types.ts
export type Design = "MINIMALIST" | "STYLISH";

export type Lang = "EN" | "ES";

export type ContentMode = "SHORT" | "LONG";

export type DownloadMode = "CURRENT" | "ALL";

export type IntlMap = {
	SUMMARY: string;
	EXPERIENCE: string;
	ACHIEVEMENTS: string;
	EDUCATION: string;
	OTHER: string;
	LANGUAGES: string;
	SKILLS: string;
	SKILLS_L1: string;
	SKILLS_L2: string;
	SKILLS_L3: string;
	PRESENT: string;
};